chore: fix changelog and markdownlint config
After the release `markdown-link-check` complained about a broken link,
which pointed to the failed 0.0.1 release which does not exist.

https://github.com/Bonial-International-GmbH/sops-check/actions/runs/12085513612/job/33702910155

I removed the link from the changelog and also removed the section for
the 0.0.1 version.

Additionally, I excluded `CHANGELOG.md` from `markdownlint` because it
complains about reoccuring headlines like `Features` in there.